requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
PyYAML>=6.0
python-dotenv>=1.0.0
xlsxwriter>=3.1.0
colorama>=0.4.6
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0